requirements.txt
mysql-connector==2.2.9
PyYAML==5.4.1
redis==3.5.3
